backend/requirements.txt
fastapi==0.101.0
uvicorn[standard]==0.22.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.2.0   # if pip can't find exactly that name/version, pip install google-generativeai